feat: refactored action url handler and revised authentication
