+++ src/handlers/auth/auth.js
@@ -87,10 +87,12 @@ export const sendEmailVerificationLink = async () => {
 export const verifyEmail = async (oobCode) => {
   try {
     await applyActionCode(auth, oobCode);
-
-    window.location.href = "/email-verification";
+    await signOut(auth);
+    toast.success("Email verified successfully");
+    return true;
   } catch (err) {
     toast.error(extractErrorMessage(err.message));
+    return false;
   }
 };
