--- src/pages/auth/emailVerification/index.jsx
@@ -1,7 +1,0 @@
-import VerifyEmail from "@/pages/auth/emailVerification/molecules/VerifyEmail";
-
-const index = () => {
-  return <VerifyEmail />;
-};
-
-export default index;
