+++ src/pages/auth/index.js
@@ -1,5 +1,6 @@
 export { default as Login } from "./login";
 export { default as SignUp } from "./signup";
-export { default as EmailVerification } from "./emailVerification";
+export { default as VerifyEmail } from "./verifyEmail";
+export { default as VerifiedEmail } from "./verifiedEmail";
 export { default as ForgotPassword } from "./forgotPassword";
 export { default as ResetPassword } from "./resetPassword";
