+++ src/pages/auth/verifiedEmail/index.jsx
@@ -1,0 +1,7 @@
+import VerifiedEmail from "@/pages/auth/verifiedEmail/molecules/VerifiedEmail";
+
+const index = () => {
+  return <VerifiedEmail />;
+};
+
+export default index;
