+++ src/pages/auth/verifyEmail/index.jsx
@@ -1,0 +1,7 @@
+import VerifyEmail from "@/pages/auth/verifyEmail/molecules/VerifyEmail";
+
+const index = () => {
+  return <VerifyEmail />;
+};
+
+export default index;
