+++ src/setup/slices/user/userSlice.js
@@ -1,5 +1,7 @@
 import { createSlice } from "@reduxjs/toolkit";
 
+import { handleEmailVerification } from "@/setup/slices/auth/authSlice";
+
 const initialState = {
   uid: "",
   email: "",
@@ -19,6 +21,11 @@ const userSlice = createSlice({
       state.photoURL = action.payload.photoURL;
       state.displayName = action.payload.displayName;
     },
+  },
+  extraReducers: (builder) => {
+    builder.addCase(handleEmailVerification.fulfilled, (state, action) => {
+      state.emailVerified = action.payload;
+    });
   },
 });
